backend/subnet_discovery/requirements.txt
neo4j
python-dotenv
ipaddress